refactor(meta): split HummockManager impl into several files.
