+++ src/meta/src/hummock/manager/compaction.rs
@@ -1,0 +1,28 @@
+// Copyright 2022 Singularity Data
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use std::collections::BTreeMap;
+
+use risingwave_hummock_sdk::{CompactionGroupId, HummockCompactionTaskId};
+use risingwave_pb::hummock::CompactTaskAssignment;
+
+use crate::hummock::compaction::CompactStatus;
+
+#[derive(Default)]
+pub(super) struct Compaction {
+    /// Compaction task that is already assigned to a compactor
+    pub compact_task_assignment: BTreeMap<HummockCompactionTaskId, CompactTaskAssignment>,
+    /// `CompactStatus` of each compaction group
+    pub compaction_statuses: BTreeMap<CompactionGroupId, CompactStatus>,
+}
